+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Version 3.0.92 (So far...) [Hero System 6e (Unofficial) v2](https://github.com/dmdorman/hero6e-foundryvtt)
 
 - Use the flash defense of the target and not the attacker for sense affecting attacks. [#1174](https://github.com/dmdorman/hero6e-foundryvtt/issues/1174)
+- Fixed TELEKINESIS damage. [#1177](https://github.com/dmdorman/hero6e-foundryvtt/issues/1177)
+
 
 ## Version 3.0.91 [Hero System 6e (Unofficial) v2](https://github.com/dmdorman/hero6e-foundryvtt)